test: use POSIX-compliant printf
The test shell script had previously used \x notation, which is
not POSIX-compliant and therefore is "undefined behavior." In this case,
the script will fail if /bin/sh is DASH or some other minimal shell.
This commit replaces the hexadecimal \x notation with \ddd octal
notation, which is strictly POSIX-compliant and will work in any POSIX
shell located at /bin/sh.